Update .golangci.yml and fix linter issues

+++ internal/shared/api/middlewares/content_length.go
@@ -16,12 +16,12 @@ func ContentLengthMiddleware(cfg *config.Config) func(http.Handler) http.Handler
 		return http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
 			if _, ok := methodsToCheck[r.Method]; ok {
 				// immediately return error if content length exceeds cfg maxContentLength size
-				if r.ContentLength > int64(cfg.Server.MaxContentLength) {
+				if r.ContentLength > cfg.Server.MaxContentLength {
 					http.Error(w, "Request Entity Too Large", http.StatusRequestEntityTooLarge)
 					return
 				}
 				// limit the size of the request body
-				r.Body = http.MaxBytesReader(w, r.Body, int64(cfg.Server.MaxContentLength))
+				r.Body = http.MaxBytesReader(w, r.Body, cfg.Server.MaxContentLength)
 			}
 			next.ServeHTTP(w, r)
 		})
